Fixing wrong page size in pagination
In Sirius-Biz pages have a default size of 50 items per page. However, in Sirius-Web the default size is 25. The page size in Sirius-Web was not set by Sirius-Biz so the pagination showed 50 items per page but calculated next and previous page items with 25. By setting the page size in Sirius-Web this should be fixed.