Fixes scaling of images with alpha channel with Java
The JPEG writer screws up the channels when an alpha channel exists.
Instead of creating a BufferedImage with an alpha channel and then
relying on the JPEG writer to deal with it, now we always make it a
RGB image and fill the background with white when drawing the original
image.