Implements a minimum size for image versions
The version string of VersionManager now supports the specification of
a minimum size. The new syntax is a comma separated string of key-value
pairs in the form of key:value. The available keys are 'size' for the
requested maximum size, and 'min' for the minimum size. The value is the
size in the format <width>x<height>.

Example: The version 'size:500x500,min:100x100' would scale an image
with 1000x800 pixels to 500x400. An image with 1000x150 would be scaled
to 500x75 and then extended with a white border to 500x100. An image
with 400x50 pixels would not be scaled but extended to 400x100.