Use Optionals for file urls
If a file was requested via the DownloadBuilder which does not exist in the storage engine, this would lead to a NPE. This is now solved with these changes. Furthermore, to improve handling of this case, an Optional is returned for the download URL instead of an String.

Tags: download, storage