Replaces an implicit entity update with an explicit query.
Updating the account from the current UserInfo is kinda
dangerous as this is a cached value. If a user performs
frequent logins (i.e. calling a some services more or
less parallel where each performs a login), this might
lead to inconsistent data. Also it updates more fields
than required (as BeforeSave handlers are executed).