Provides the "events" part of the analytics framework.
This is the lowest stage of the metrics and analytics framework and
responsible for collecting raw events into an analytics database
which is Clickhouse in our case. To make processing as efficient
as possible and to also not slow down the application itself,
events are collected and queue in an internal buffer and then
batch-inserted into the database.