Prevent a file being replaced by the same file
If a file is uploaded in the storage engine and a md5 hash is provided, it can be checked whether the file has been changed with the already stored md5 hash. In this case it is not necessary to overwrite the file and delete all of its versions.

Tags: storage, upload