Perform reindex async, adjust dynamic mappings
As the REST client timeouts after 30s, we just perform the request async
and inform the user to check the task API.
Before performing the reindex, we create the index with dynamic mapping
mode to "false". This is done so that the process is not aborted by
possible legacy fields in documents.
After moving the active alias, the dynamic mapping mode is shifted to
"strict" which will cause an error if the application tries to index
fields which are not part of the mapping.