Changes the wording of the ip/scope parameter to be more generic
This is only in the default implemention the IP and can be used in a much broader sense for grouping multiple calls that should be counted towards the same limit.
For example this can and will be used to limit all API calls of the same tenant. The new name should now reflect this better.

Fixes: SIRI-55