Permits to ignore i5 data conversion errors.
This is required if the "Java side" is updated before the i5 side as most probably the
left over byte buffer is uninitialized and will crash data conversions. The application
has to ensure itself that this data is never used up until it is populated correctly.

Also this feature is turned off by default for obvious reasons.