Splits Isenguard and the Firewall impl into two.
This way we can always register the Isenguard service even if the
framework is disabled. This is required, as some UI components as
(w.rateLimitBox) need them in scope.

The Firewall is only enabled and registered, if the framework is active,
therefore the observable state result will be the same.