Properly cleans the permissions / flags being used to identify the sys admin / sys tenant.
For Users we now grant two flags (can be checked via UserInfo.hasPermission):
flag-system-tenant-member - granted to all users which belong to the system tenant
flag-system-administrator - granted to administrators of the system tenant
(administrator = have permission 'permission-manage-system')

For Tenants we now grant 'flag-system-tenant' (can be checked via Tenant.hasPermission).
This can be used in 'required-permissions-for-permission' mappings to filter specific roles
or permissions on system tenant users.