Add facet items in the addFacet callback
Because we need to add the items after we set the selected value to the facet, because the value should already be present when adding items so it can be determined there, whether the item is active or not. The value of the facet will be set, in the addFacet method, before the callback is called.