Allows for a job parameter based on a list of priorized parts
This allows to select from a list of classes that implement and register a common superclass when starting a job.
The list of provided select options is sorted by the priority of the parts.
As a label for the single select toString of the classes is called.
The job is provided with the registered implementation object of the selected part.