We must not ignore handled exceptions here.
First of all we're not expecting any exception here, as "invokeChecks" is set to false
(the checks are executed when recording the event).

Second, if a database error occurs (like clickhouse being down), we'll
also receive a handled exception from "insert" and should abort right here. This
will let the loop sleep for another 5 minutes and will give clickhouse enough time
to startup...