Prevents the user from entering a non existent Entity as parameter
The inputfield allows to enter values that are not present in the autocomplete suggestions which currently leads to a null parameter value. Which is fine for mandatory parameters, as this catches the invalid input. But for optional parameters this results in an error while executing the job and not in an error message when trying to start the job - which is desired.

Fixes: OX-5335