Make the parsing of dates strict
This detects wrong dates like the 31.09.2019 (September only has 30 days).

The trade off to this solution is, that years have to be declared with 'u' in pattern instead of 'y'.
The other option one could achieve strict parsing with is the 'SimpleDateFormat'-class, but it comes with its own set of problems, also it is legacy java.

Additional reading from StackOverflow:

DateTimeFormatter:
`uuuu` versus `yyyy`: https://stackoverflow.com/questions/41177442/uuuu-versus-yyyy-in-datetimeformatter-formatting-pattern-codes-in-java
Issues with `yyyy` isntead of `uuuu`: https://stackoverflow.com/questions/41103603/issue-with-datetimeparseexception-when-using-strict-resolver-style

SimpleDateFormat:
Ignores additional/missing characters in pattern: https://stackoverflow.com/questions/16014488/simpledateformat-parse-ignores-the-number-of-characters-in-pattern
Make SimpleDateFormat more strict (see comments of top answer): https://stackoverflow.com/questions/13088140/java-how-to-parse-a-date-strictly