Enforces some constraints before actually deleting or updating entities.
The main reason for doing this is, that some import handlers will yield
readonly entities from their parent tenants. These can be used as references
etc. but must neither be deleted or updated.

Everything else should most probably be handled during the load or within
a beforeSave or beforeDelete handler.