Simplifies the way children are resolved.
There is no real benefit in using an Optional, as the find must always return a VirtualFile
(which might or might not exist). Otherwise an error condition is present and an Exception
should be thrown anyway.

This changes the behavior so that a placeholder file is created and marked as non existing
instead of (crashing) with an empty optional.