Simplifies and unifies the way code lists are scanned and cached.
We not only and always cache the pair of value/additional value. Additionally
we consistently provide a single API (tryGetValues) which permits to distinguish
between "code" does not exist and will not be created or code is available or would
be created (due to auto fill). All other APIs (hasValue / getRequiredValue / verifyValue)
use this internal method.