Uses createNowOrUpdateInBatch as default way of updating entities.
Otherwise "unique" checks might fail (at the database level) as the inserts are queued in the batch
and therefore not enforced by the framework. This adds too much trouble for outweight the performance improvements
of a batched insert.

If an entity is known to not have any unique constraints / such conflicts - which is probably very rare - this method
can be overwritten and changed back to a batched insert.