Drops all automatic data checks and provides boilerplate methods instead.
The automatic checks where sometime too strict, sometimes too lose. Also within a composite
one cannot access the "isChanged" detection of the underlying entity, which is a good way to
prevent tons of unnecessary checks in the first place.

Each entity can now itself implement a BeforeSave method and/or an OnValidate method which
calls the appropriate checks of the composites.