Enhances the Importer to receive the job factory name used to initialize it
This opens the door for multiple ImportHandlers to be defined for the same target Entity.
Although in most cases only a single handler would exist for a given entity, there are some needs where a custom handler must be created for specific import or export tasks. In this case the handler can validate if its being used under the context of a specific factory in its accept method by using context.getImporter().getFactoryName().

Note that when using this feature, the default mappings and dictionary must be defined exclusively by the job factory, making the handler sort of exclusive to the factories.