Improves detection of broken control connections.
In some case we observed a JDK bug where the Socket
itself felt "connected" but the SocketOutputStream reported
a "broken pipe" when trying to access it.

The developers of the Apache FTP client also observed this
behavior and provided us with a more elaborate way of checking
of the connection is _really_ available.

(cherry picked from commit 64d255941f9539fa1d177babb62e7472161bd96e)