Adds some retry magic for the FTPUplink.
As we detected some IO exceptions when talking to some FTP
uplink servers, we added this simple retry patterns to eliminate
the "Broken Pipe" errors. Although we check the validity of the
underlying socket connection quite intensively, we still observe
SocketExceptions for broken pipes.

Hopefully a simple retry/reconnect solves this problem.