Also test connectors being returned into the pool.
If a retry-able io operation crashes, we attempt a retry
but we also forcefully close the connector we used.

This will invalidate the object. By enforcing a validation
before actually returning the object into the pool we
can save some overhead. Note that this shouldn't
have any semantic side-effects as the connector would
be tested before its use anyway - therefore this is really
just a shortcut.