Provides various ways of spelling email.
Note that the EMail capitalization isn't necessary at all.
We always compare lowercase. This is simply a cheap
trick to not receive a duplication warning as the translation for
email is still email in most languages. However, someone
might change it one day (e.g. to E-Mail) - we still want to
be able to import "email" columns. Therefore this redundancy
is placed here.