Revert "Set length of tennantId and userId to 26"
This reverts commit 6cdf803d

As the user_id for users created by MongoTenantUserManager has a length of 43 this composite did not work.
This composite should provide a more loose length for ids and not restrict it as its has no connection to a specific UserManager Implementation and therefore should handle a variety of sensible lengths for ids.