Refactors the tokenizing of searchable prefixes
Introducing a BiConsumer as base interface for Transformers, which shall implement the logic to deal with the property type and tokenize its items by invoking the given consumer.
Same as before, property.getValue is used when no transformer exist.

Note that addContent has been removed as the logic in this method should be implemented as Transformers.

Fixes: SIRI-171