Provides a UI component to update the value of a BlobHardRef.
This is done by performing an AJAX upload into a temporary blob.
Once the entity containing the BlobHardRef is persisted,
the blob is marked as non-temporary. Otherwise, the blob is
automatically deleted after a timeout.

The controller provided here supports the actual AJAX call.

Note that the UI is very rudimentary and more or less a copy of
storedObjectUploader.html.pasta. This needs some love+work
in the near future.

Fixes: SIRI-96