Stores the raw user input instead of the already transformed values.
This is because in JobFactory.startInBackground the parameters are checked
and transformed again. We to not expect that the output of a parameter
is "re-transformable", it only needs to be resolvable by the parameter
itself.

We also do not want to bypass the checkAndTransform in
JobFactory.startInBackground as for JobPresets, to can only store the raw
input and therefore have to rely on the check later on. (This is because
JobPresets are expected to be incomplete and therefore an upfront check
would lead to false positives).

The underlying bug (SIRI-195) was that the LocalDateParameter parsed
all kinds of user input and then returned an ISO representation which
again was not parsable by the underlying AdvancedDateParser itself.

Fixes: SIRI-195