Disabled batch updates/inserts by default.
We ran into several problems with MariaDB when performing concurrent
batch updates. This happens, as during a batch run, sections of the primary
key get locked which might lead to a deadlock within MariaDB if concurrent
batches update the same table (which isn't that unlikely).

As this would/will abort one batch update/transaction and therefore would
lead to dataloss / an abort of an import, we rather play safe (by default),
and always succeed.