Centralizes alias handling.
We now always add the name and label of a field as an alias
into the import handler. This was previously performed in the
BaseImportHandler when also loading the aliases from the
config.

However, this doesn't work for manually created aliases,
therefore this was moved into the dictionary itself.

We also keep the aliases in an non-normalized state for reporting
them to the user in a proper way.

Finally, we only report a fatal alias error, if the alias points to another
field. Everything else is only reported if a config error is present in the
system configuration.