Improves the stability of stream based uploads to S3.
If a stream is uploaded to S3, we now check its size. If it is
unknown or larger than 8 MB, we force a multipart upload.

If method is way more robust than performing a single upload,
as it can load all contents of a chunk in memory and then perform
an upload with the possibility of performing a retry in case of a transient
error signaled by s3. Also, we set the buffer for the remaining stream based
uploads to 8 MB so that these also fully fit into the local buffer and provide
the same retry capabilities.