Keeps the search contents in the _source field.
Otherwise we'd had to re-tokenized the data
in every update, which isn't guaranteed by the
framework.

Therefore this would open up a bunch of dataloss
problems when using this entity in certain circumstances.
Also, loading the field isn't that expensive anyway as it should
be quite small for most entities.