Removes the ID as default exported field.
However, if requested as column, it still can be exported. Keeping the ID in every export is
kind of dangerous as if present, this is the preferred way of synchronizing entities. Therefore
this prohibits and export and then import on different systems (as the IDs will most probably
not match).