BREAKING: Splits the user-management permissions.
We now distinguish between "managing users of any tenant" and "managing
users of the system tenant" as this is entirely another game.

Previously, every user of the system tenant with the role "user-administrator"
was system administrator. We now introduced now a extra role and also an
extra permission which controls if users of the system tenant may be
updated.

To achieve backwards compatibility, add this to the application.conf:

        "flag-system-tenant-member+user-administrator" {
            priority = 10
	    system-administrator = true
        }