Fix an issue creating files in the storage
The find or create method was skipping the file creation if a previous find was issue which did not find any file (leaving the cache with null for the desired key).

This caused subsequent calls to the cache to deliver null instead of forcedly call the valuecomputer in order to create a new file.

Fixes: OX-5973