Executes cache manipulation calls synchronous for coherent caches.
After invoking cache.clear (etc.) we expect the cache to be actually
cleared on the local node. However, as the Interconnect distributes
the update via Redis there is actually a race-condition built-in.

We now modify the cache locally and Ignore the Interconnect message
we sent out, to ensure proper and consistent behavior.