Always makes the Metrics instance nullable.
This is for the rare occasion where metric computers exist but metrics collection is
to be suppressed by disabling this framework. In this case the schedulers won't be
active, so an NPE is not to be expected, however, the framework will still try
to initialize all parts and might report a startup warning - these warnings are suppressed
by this change.