Refactors the extract archive job.
By default we now overwrite files if a change is detected (SIRI-279) and also permit to delete
the archive once all files are extracted (SIRI-276).

We also reduced the logging quite a bit (which is still available in debug mode). Also we now
require a destination to be given and no longer use the parent directory of the archive
as fallback (as this provides quite a potential for trouble).