Improves the replication performance by fine-tuning the queue config.
We now do not wait until the queue runs completely empty before we attempt
to refill it, but rather try to keep a certain base level of batches in the queue
(as long as work is available).

Additionally we increased the batch size and the refill frequency to adjust the
performance for high load conditions.