Adds some convenience methods to perform tasks as system tenant.
Some APIs require a tenant to be present. When executing background tasks
or work which is independent from the current user, this helpers can be used
to pretend that the current task is somehow executed by the system tenant.

Note that this will only satisfy requests which concern the tenant id or name.
The generated user id and name is constant and doesn't reflect a real user
in the database.