Doesn't specify a default port when creating an AWS client.
Previously we explicitly added the "443" port to all generated
uris. This might break some HTTP clients, as the signer within
the SDK generates signatures where the host header strips
off the default port, even if it is contained in the uri. As this
also doesn't provide any benefit, we now skip adding the port
as long as it is the default port.