Fix the vfs modal when querying big directories
The JSON response of /fs/list is limited to 32k characters, which
equates to around 150 files. Anything bigger than this leads to a broken
modal. The proper fix for this would be to introduce paging and a search
bar. As a quick fix we now limit the number of files to 100.