Prevents error when transforming force failed query to page
Force failing a query can be done to simplify logic that should not present a query result in some conditions. Facets can not be executed on a failed query and doing so results in an exception. When using the PageHelper to transforming a query into a Page for displaying it in a template we therefore should skip executing the facets when the query is marked as force failed.

Fixes: OX-6685