Support UTF-8 file names in FTP uplinks
Some servers want to speak UTF-8 with us. To support this we now enable
the autodetection, which checks if the server lists UTF-8 in it's
features. In this case we set our control connection to UTF-8. If the
server doesn't support the FEAT command before authentication, we have
no way of knowing the encoding. In these cases the encoding can be
configured for FTP uplinks.

Fixes: SE-10483