Skips redundant positive audit log entries.
If a positive audit log entry re-occurs on the same day,
we ignore it. This is quite common for service calls which
might perform a login and might be quite frequent. As long
as these are all positive, contain the same user and IP, these
is actually no point in logging many identical events. The calls
itself will most probably be traced by another Event anyway.