BREAKING: Changes the index settings for AuditLogEntries for deduplication.
The "message" field actually contains just the NLS key used to
translate the message. Therefore we can safely index this field
(which is required to de-duplicate positive entries).

Note that due to an index settings change, one must use the
"Re-Index Elasticseach Entity " and " Move Elasticsearch Index Alias"
jobs to fix this.