Ensures the login is recorded everytime findUserByCredentials is called
This method is sometimes used in other authenticating places (e.g. the FTP or SSH bridge) where previously a authentication would create an audit log entry but would not count towards the last login / number of logins of the account. This is now coupled more directly by moving the recordLogin call within the findUserByCredentials function where also the audit logs are written.

Fixes: SIRI-342