Introduces the DelegateJournalData
This class should be used instead of JournalData when there is a need to record changes under another entity (usually the parent entity), like recording changes to product's features under the product itself instead of logging such changes directly at the child entity.

It can be constructed using an entity ref (of the entity which shall own the record) or using the id + type.

Fixes: SIRI-298