Fixes a race condition were the user cache contains stale data.
After an external login, we use a direct database update to remember
the login. This will not flush the cache within the TenantUserManager.

However, after an external login, we should flush this cache as the
"lastExternalLogin" field has been updated. Otherwise, we might be
redirected to the login page again as the user seems still outdated.