Handles the routing after /tenants/select/main or /user-accounts/select/main.
We now always stay on the selection page. HOWEVER, if
/user-accounts/select/main is invoked but only the tenant and not the
user itself is switched, we redirect to /tenants/select/main which will
clear the tenant.

This way, we can use /user-accounts/select/main to clear the user OR
the tenant with a single link - this is used by a badge which shows the
user and tenant name if one has switched either of both.