Removes a fallback routing which is now handled via the DefaultDispatcher.
As we want to permit applications to handle "/" in any way, we cannot
provide a controller for it - not even a low ranked one, as there might
be a dispatcher which actually handles the request but runs after the
ControllerDispatcher.

We therefore only handle /admin and /dashboard but let the
DefaultDispatcher in sirius-web attempt to "redirect" (restart the
pipeline) from / to /admin to let this controller handle the request
in favor of sending a 404.