No longer always fills the total items with the current items
If we don't know the total size, we don't want to supply a total item count to the page. This bug was introduced while trying to fix another bug, where fetchTotalItems wasn't applied for non-full pages that weren't at start=0.
This change keeps this fix so total items will be filled when
a) withTotalCount is enforced
b) we are at start=0 and we don't have any more items
In all other cases the totalItems will not be supplied to the Page at all.