Changes the default cache-behavior of URLBuilder.
We previously delayed the lookup of the physical key as much as
possible.

We therefore generated virtual delivery URLs which delay the lookup, but
are inherently less cacheable. This has now been changed and unless
URLBuilder.delayResolve() is invoked, we try to generate a physical
delivery URL if possible. The lookup is backed by a cache anyway and
the benefit of having an eternally cacheable media file overweights
cost of the lookup anyway.