Greatly simplifies loading values into a MultiLanguageString.
We previously invested quite some effort to ensure that only proper
language keys ended up in the MLS. However, this is quite an overkill,
as we only need to check and sanitize user input (when reading from
a web request).

We now ensure that no malicious input is present (only valid language
codes end up in the map) - everything else should be fine anyway.

Also a legacy hack has been removed as it is no longer required.