Changes the way the validLanguages of an ML-String are determined.
Previously an explicit list was passed in. However, in many cases, the
underlying entity is constructed, but the ML-String isn't accessed
at all or just in a way which doesn't concern the language list.

Therefore the valid languages are now passed in as Supplier and
lazily evaluated. Also, instead of an explicit supplier, one now can
supply a target scope type and the framework will use the display
languages of the given scope as list of languages.

If neither of both is given, we use the display languages of the
default scope.

Fixes: SIRI-422