Moves hint message constants and introduce a new hint
as most of the logic is solely handled in ProcessLog, makes no sense to keep these constants defined in ProcessContext.

In addition, we introduce a new hint: HINT_MESSAGE_KEY, which is supposed to be used when providing translatable keys. They will be automatically appended with the `$` (dollar sign) and therefore smart translated. Use HINT_MESSAGE_TYPE for raw text which demands no translation

Fixes: OX-7331