Improves recording of failed conversion attempts.
For a failed attempt, we still update the duration counters. We also
use a lower level of database API, as technically, we update an
outdated entity. The given variant has internally still
"queuedForConversion: false", as the update to true, also happens at
a lower level and after that, no entity update is performed. Now, if
we try to toggle the flag back to false, the framework detects, that
this is a NOOP and doesn't perform an update at all - which leaves
the variant hanging at "queuedForConversion" in the case of an error.

Using the low-level API, this problem doesn't occur.