Fixes url after following wrongly encoded redirect headers
By standard, the redirect header should contain the url in ISO-8859-1;
But most browsers and a lot of other clients also support UTF-8 so some servers will use UTF-8. When UTF-8 is used, the wrongly interpreted symbols will always start with the character 'Ã'. As the chance for this symbol in uppercase to be used in a correct url is pretty low, and at this point of the code we are at our last resort anyway, we will re interpret the String as UTF-8 if we come across this character.