Bugfix: Only applies empty checkbox marker for boolean properties
This fixes a bug where lists filled by checkboxes could not be properly be emptied.

We have to check for a BooleanProperty here, because checkboxes might also be used in templates with different values and the same name to let the user fill/unfill a ListProperty or BaseEntityRefListProperty - and these templates will also send markers for the given propertyName, so this lead to these lists be filled with a single entry 'false' when they were supposed to be empty.

Fixes: OX-7503