Limits the checks if a virtual file should be updated from an URL.
If the file exists and has been modified (or at least touched) on the
same day, we don't need to call the given URL (even if we abort early
using an If-Modified-Since header - as these are often not supported).

If a download is to be forced, the mode "ALWAYS_FETCH" can be used.
Note that in order to make this work properly, we "touch" an existing
file if the download is skipped due to a NOT_MODIFIED response -
otherwise these files will be re-attempted over and over again.

Fixes: OX-7437