Changes the resolve order for filenames from URLs and makes a helper method public.
If we try to extract a filename/filepath from a given URL, we now FIRST
check the URL path, and only if this doesn't yield a result, we check
the query parameters. Note that it all boils down to have a proper
"fileExtensionVerifier" which selects the valid path or parameter.