Adds an optional file extension filter for file picker modal
In wondergem, this will only limit the files displayed.
In tycho, this will also restrict the upload box to the given file types.
We use a comma seperated list of file extensions, each prefixed with '.' so we can easily feed the list directly into the dropzone library, and we can easily convert the string into a list(stream) on the server.

Fixes: SIRI-469