Handles exception for non-rfc-urls in location header
currently non-rfc-conforming URLs (e.g. containing unescaped spaces) in the location header when redirecting will break the http client and abort with an IllegalArgumentException. We need to find a better solution for this, but temporarily we just catch it here and handle the exception, so we can provide appropiate logs and we don't abort completly.

Fixes: SIRI-479