Normalizes to Unicode NFC Form when sanitizing a path.
This effectively means that all lookups and therefore each new blob creation is normalized to the NFC form. This also means that existing NFD forms an no longer be found, even when querying for exactly the NFD form (as it will get normalized). Therefore all existing DB objects (MongoDirectory,Mongoblob, SQlBlob...) need to be normalized to NFC. An appropriate job for this task will be provided in the next commit.

Fixes: SIRI-491