Fixes Lookup of Blobs via Filename.
SQLBlob & MongoBlob and SQLDirectory & MongoDirectory sanitize the filename and additionally lowercase it for the normalized filename in a @BeforeSave method.
Thus, when looking up a blob via its filename, it also has to be sanitized before the lookup, as no blob will be found otherwise in case the filename contains characters that are removed or replaced when sanitizing.

This issue occurred in a file export job where the file to be exported included the time that was formatted with colons (which are replaced by underscores during the sanitizing). This lead to the uniqueness check failing in `BasicBlobStorageSpace#isAttachedBlobUnique` and subsequently failing the job.

- Fixes: SIRI-519