Exposes methods to only log an exception
The current logic handles errors thrown nicely logging them under the current context, but this is not sufficient.
Many times we need to log a warning, respecting the current context but carry on with processing.
This way we expose 2 methods which can be used to simply log the HandledException given.
This way we also profit from the hints attached to the message, especially the ones used to limit the number of incidences to be logged.

Fixes: OX-8105