Replaces custom global metric computers by ones bound to an actual entity type.
This has been done for two reasons. First, it simplifies the framework,
as we don't have extra schedulers to maintain. Second, it permits to
properly order the execution of computers (via levels). As we commonly
use the global computers to aggregate data computed by other (entity
based) computers, this is frequently required to compute proper results.

(Note that the level of a computer is only honored within the same
scheduler and not across different ones.)