Accounts for race-conditions when scheduling the next level of analytics tasks.
An analytical scheduler can group its tasks into distinct levels. We
guarantee, that a successive level is only executed, if all tasks of the
preceding level have been completed (as the computers might rely on the
results).

However, as the framework probably runs in many threads on many nodes,
some batches might have been started in parallel to the last batch
(which is in charge of scheduling the next level). Therefore, we use a
named region to ensure that no other parallel computations run, before
we schedule the next level (with a safety timeout of 30min). Note, that
we don't have to check the queue as we know we were the last batch so
we only have to account for preceding ones which might run in parallel.