Specifies a marker interface which permits to directly query for child files.
This approach was chosen (in favor of extending the listXXX methods), as
we most probably only need this functionality in the L3Uplink, where we
can provide elaborate filter facets as we access the underlying database.

However, all other implementations of the VFS don't have this capability,
so the leave the interface as they are and provide this extension instead.

Fixes: SIRI-587