Moves Interactive Jobs Pages to Tycho.
As the page is now wider, the additional metrics block is moved to its own column on the right in order to simplify the logic in the template as well as use the additional space instead of cramming it below the job parameters.
This change also comes with using the existing execute and back buttons used in other jobs instead of splitting them up like before. In case no parameters can be set, a back button will be inserted below the graph.

Note, that the column width of the graph in the middle is defined manually instead of relying on the bootstrap framework by simply using "col-md", as some graphs like the bar chart do not calculate a correct width otherwise.

- Fixes: SIRI-615