Migrates and overhauls the overview page for scheduled jobs.
Instead of tables, cards are now displayed.
Each card represents one entry and lists some of the details.
To keep the cards short and lower the amount of useless information, the default 'asterisk' value for execution parameters won't be displayed.

Additionally, some "Ausführungen" is now a proper NLS key with translations and neighboring english translations are fixed.

- Fixes: SIRI-616