Displays job icons correctly in wondergem
We still use the w:jobs templates which links pages to capable jobs, and there the icons have gone mad, especially when they no longer match the naming, eg: fa-file -> fa fa-file works, but fa-file-excel-o is no longer found.

Note that we cannot mark this method as deprecated as biz would refuse to compile and will remove this with a future ticket.

Fixes: OX-9090