Don't override security config arrays from other conf files
When merging multiple config files, HOCON does merge objects, but not arrays. When you specify an array for e.g. security.roles in another component-<sth>.conf file, it is randomly chosen which one is loaded first. So: If a value was already loaded, we should merge it, instead of overriding it.