Adds separate metric for number of delayed replication tasks
This metric is not really necessary as those are already represented by the difference between the total number of tasks and the number of executable tasks.
But it helps to further explain that a high number of total tasks may be expected as delete tasks are delayed for a set number of days. These are not actual replications (like backing up an uploaded file) but rather deletions of the backed up file after a file is deleted in the main storage space.

Fixes: SIRI-672