Increases the size of the 'name' column for Facts.
In most cases, 50 characters are enough as the field mostly contains fixed size strings for a given entity type and id. In special cases however, an entity might have multiple statistics of the same type, but for specific "sub categories".
The name then consists of type + identifier of the "sub category". If the sub category doesn't have an ID and is instead only identifiable by a label or name of some sort, the total name may exceed the 50 character limit.

- Fixes: SIRI-678