Don't swallow exceptions in the completion handling of watchable stream s
Currently, WatchableInputStream and WatchableOutputStream use a future for the completion handling. The futures are fulfilled, once the stream is closed. Unfortunately, our Futures swallow exceptions that happen during completion handling.

The FTPClient that FTPUplink uses, mandates to check the reply code after data transmissions. If the code is an error code, the transmission wasn't successful. In order to have any chance of handling these transmission errors, we need to throw an exception.

This commit replaces each future with two simple completion handlers. This change is breaking as any exception that was only logged away before, is now propagated.

Fixes: SIRI-682